requirements/base.txt
asgiref==3.2.10
pytz==2021.1
sqlparse==0.4.1
influxdb==5.3.1
celery[redis]==4.4.2
redis==3.5.3
requests==2.25.1
psycopg2==2.8.4
dj-static==0.0.6
gunicorn==20.0.4
pandas==1.4.2
Pillow==8.2.0
openpyxl==3.0.9
pygbif==0.6.3
whitenoise==6.6.0

# DJANGO------------------------------------------------------------------
Django==3.1.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-environ==0.4.5
django-material-admin==1.8.5
django-redis==5.0.0
django-filter==2.4.0
django-phonenumber-field[phonenumbers]==5.2.0
django-imagekit==4.0.2
django-import-export==2.7.1

# DJANGO REST FRAMEWORK---------------------------------------------------
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
django-rest-swagger==2.2.0
drf-yasg2==1.19.4